# apiextensions: Benchmark escaping in SchemaHas and pool Schemas