@apelisse 2023-01-30
/triage accepted
/sig wg-api-expresison
/cc @lavalamp @alexzielenski @Jefftree 
/priority important-soon
I don't see how that's an API change?
/remove-kind api-change